+++ app/src/main/java/com/jaguarsoft/mac/favouriteplaces/backend_sdk/model/Coordinates.java
@@ -1,6 +1,6 @@
 package com.jaguarsoft.mac.favouriteplaces.backend_sdk.model;
 
 public class Coordinates {
-    public float longitude;
-    public float latitude;
+    public double longitude;
+    public double latitude;
 }
